Merging 1.11.2 back to Master

+++ src/main/java/brightspark/stem/message/MessageRecipeReply.java
@@ -3,7 +3,6 @@ package brightspark.stem.message;
 import brightspark.stem.recipe.ClientRecipeCache;
 import brightspark.stem.util.CommonUtils;
 import io.netty.buffer.ByteBuf;
-import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraftforge.fml.common.network.simpleimpl.IMessage;
 import net.minecraftforge.fml.common.network.simpleimpl.IMessageHandler;
@@ -12,11 +11,11 @@ import net.minecraftforge.fml.common.network.simpleimpl.MessageContext;
 public class MessageRecipeReply implements IMessage
 {
     public ItemStack recipeStack;
-    public int recipeFluid;
+    public long recipeFluid;
 
     public MessageRecipeReply() {}
 
-    public MessageRecipeReply(ItemStack stack, int fluidAmount)
+    public MessageRecipeReply(ItemStack stack, long fluidAmount)
     {
         recipeStack = stack.copy();
         recipeFluid = fluidAmount;
@@ -26,14 +25,14 @@ public class MessageRecipeReply implements IMessage
     public void fromBytes(ByteBuf buf)
     {
         recipeStack = CommonUtils.readStackFromBuf(buf);
-        recipeFluid = buf.readInt();
+        recipeFluid = buf.readLong();
     }
 
     @Override
     public void toBytes(ByteBuf buf)
     {
         CommonUtils.writeStackToBuf(buf, recipeStack);
-        buf.writeInt(recipeFluid);
+        buf.writeLong(recipeFluid);
     }
 
     public static class Handler implements IMessageHandler<MessageRecipeReply, IMessage>
